Refactor createNew odsp flow to prepare for a new partition-based flow
