+++ packages/drivers/odsp-driver/src/test/createNewUtilsTests.spec.ts
@@ -89,6 +89,7 @@ describe("Create New Utils Tests", () => {
             },
             new TelemetryNullLogger());
         newFileParams = {
+            type: 'New',
             driveId,
             siteUrl,
             filePath,
