+++ packages/drivers/odsp-driver/src/test/getVersions.spec.ts
@@ -39,6 +39,7 @@ describe("Tests for snapshot fetch", () => {
     const resolvedUrl = ({ siteUrl, driveId, itemId, odspResolvedUrl: true } as any) as IOdspResolvedUrl;
 
     const newFileParams: INewFileInfo = {
+        type: 'New',
         driveId,
         siteUrl: "https://www.localhost.xxx",
         filePath,
